# Improve rolling bloom filter performance and benchmark