@dcousens 2016-04-26
light utACK 67d44f8,  didn't verify masking operations