@sipa 2016-04-28
Addressed @dcousens's nits.