@rebroad 2016-12-21
I'd like to understand this code - where do I start?